Popeye/automation.py
import nmap
import requests

class AutomationManager:
    def __init__(self):
        self.commands = ["scan network", "exploit vulnerability", "generate report"]

    def execute(self, command):
        if command == "scan network":
            self.scan_network()
        elif command == "exploit vulnerability":
            self.exploit_vulnerability()
        elif command == "generate report":
            self.generate_report()

    def scan_network(self):
        print("Scanning network using Nmap...")
        nm = nmap.PortScanner()
        nm.scan('127.0.0.1', '22-443')
        print(nm.all_hosts())

    def exploit_vulnerability(self):
        print("Exploiting vulnerability using Metasploit...")
        # Add Metasploit exploitation logic here

    def generate_report(self):
        print("Generating report...")
        # Add report generation logic here