Popeye/requirements.txt
speechrecognition
pyaudio
shodan
virustotal-api
requests
beautifulsoup4
python-nmap
recon-ng
metasploit
haveibeenpwned
toggl
wave
trello
youtube-dl
unsplash
pexels
openai
coursera
khanacademy
google-api-python-client
upwork
fiverr
python-dotenv
pip-audit
bandit
pytest
twilio
python-telegram-bot